+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Features
 - User impersonation (#188, PLUM Sprint 230509)
 - Configurable client session expiration (#204, PLUM Sprint 230509)
+- Additional TLS options in LDAP provider (#208, PLUM Sprint 230519)
 
 ---